# perf: parse Loki labels in protobuf write path