--- a/src/servers/src/http/loki.rs
+++ b/src/servers/src/http/loki.rs
@@ -27,7 +27,7 @@ use axum::{Extension, TypedHeader};
 use bytes::Bytes;
 use common_query::prelude::GREPTIME_TIMESTAMP;
 use common_query::{Output, OutputData};
-use common_telemetry::warn;
+use common_telemetry::{error, warn};
 use hashbrown::HashMap;
 use lazy_static::lazy_static;
 use loki_api::prost_types::Timestamp;
@@ -257,7 +257,7 @@ async fn handle_pb_req(
     for stream in req.streams {
         let labels = parse_loki_labels(&stream.labels)
             .inspect_err(|e| {
-                warn!("failed to parse loki labels: {}", e);
+                error!(e; "failed to parse loki labels");
             })
             .ok();
 
@@ -319,7 +319,8 @@ pub fn parse_loki_labels(labels: &str) -> Result<BTreeMap<String, String>> {
 
         // parse value
         let qs = quoted_string::parse::<TestSpec>(labels)
-            .map_err(|_| {
+            .map_err(|e| {
+                error!(e.1; "failed to parse quoted string");
                 InvalidLokiLabelsSnafu {
                     msg: format!("failed to parse quoted string near: {}", labels),
                 }
@@ -329,7 +330,8 @@ pub fn parse_loki_labels(labels: &str) -> Result<BTreeMap<String, String>> {
 
         labels = &labels[qs.len()..];
 
-        let value = quoted_string::to_content::<TestSpec>(qs).map_err(|_| {
+        let value = quoted_string::to_content::<TestSpec>(qs).map_err(|e| {
+            error!(e; "failed to unquote the string");
             InvalidLokiLabelsSnafu {
                 msg: format!("failed to unquote the string: {}", qs),
             }