I hereby agree to the terms of the [GreptimeDB CLA](https://github.com/GreptimeTeam/.github/blob/main/CLA.md).

## Refer to a related PR or issue link (optional)
#5287 

## What's changed and what's your intention?

Parse Loki labels directly in protobuf write path instead of converting and decoding using JSON5, roughly 10x better performance-wise while ensuring correctness.

```
loki_labels/json5       time:   [3.9928 µs 3.9939 µs 3.9950 µs]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high severe
loki_labels/hand_parse  time:   [392.19 ns 392.54 ns 392.93 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```

## PR Checklist
Please convert it to a draft if some of the following conditions are not met.

- [ ] I have written the necessary rustdoc comments.
- [x] I have added the necessary unit tests and integration tests.
- [ ] This PR requires documentation updates.
- [ ] API changes are backward compatible.
- [ ] Schema or data changes are backward compatible.